fmbttizen: take screenshot on weston

+++ utils/fmbttizen.py
@@ -30,7 +30,7 @@ Then save the lock on the lockscreen and as "lock.png". Install
 shutter and run in shell:
 
 display /tmp/lockscreen.png &
-shutter -s --exit_after_capture -o lock.png
+shutter -s -e -o lock.png
 
 
 Example 2: Open the lock screen, launch Settings
@@ -56,6 +56,7 @@ import base64
 import cPickle
 import commands
 import math
+import shlex
 import subprocess
 import os
 import Queue
@@ -120,22 +121,33 @@ def _fileToQueue(f, outQueue):
     f.close()
 
 class Device(fmbtgti.GUITestInterface):
-    def __init__(self, serialNumber=None, debugAgentFile=None):
-        """
-        Parameters:
+    def __init__(self, serialNumber=None, loginCommand=None, debugAgentFile=None):
+        """Parameters:
 
           serialNumber (string, optional)
                   the serial number of the device to be connected.
                   The default is the first device in "sdb devices"
                   list.
 
+          loginCommand (string, optional)
+                  Shell command for establishing SSH connection to the
+                  device. Example:
+
+                  loginCommand = "/usr/bin/ssh tizen@192.168.99.20"
+
+                  If loginCommand is given, sdb will not be used.
+
           debugAgentFile (file-like object, optional)
                   record communication with the fMBT Tizen agent to
                   given file. The default is None: communication is
                   not recorded.
+
         """
         fmbtgti.GUITestInterface.__init__(self)
-        self.setConnection(TizenDeviceConnection(serialNumber=serialNumber, debugAgentFile=debugAgentFile))
+        self.setConnection(
+            TizenDeviceConnection(serialNumber=serialNumber,
+                                  loginCommand=loginCommand,
+                                  debugAgentFile=debugAgentFile))
         self._serialNumber = self._conn._serialNumber
 
     def close(self):
@@ -241,9 +253,10 @@ class Device(fmbtgti.GUITestInterface):
         finger2endX = int(x + math.cos(math.radians(finger2Dir)) * endDistanceInPixels)
         finger2endY = int(y + math.sin(math.radians(finger2Dir)) * endDistanceInPixels)
 
-        return self._conn.sendMtLinearGesture([[(finger1startX, finger1startY), (finger1endX, finger1endY)],
-                                               [(finger2startX, finger2startY), (finger2endX, finger2endY)]],
-                                              duration, movePoints, sleepBeforeMove, sleepAfterMove)
+        return self._conn.sendMtLinearGesture(
+            [[(finger1startX, finger1startY), (finger1endX, finger1endY)],
+             [(finger2startX, finger2startY), (finger2endX, finger2endY)]],
+            duration, movePoints, sleepBeforeMove, sleepAfterMove)
 
     def pinchBitmap(self, bitmap, startDistance, endDistance,
                     **pinchAndOirArgs):
@@ -478,10 +491,24 @@ class TizenDeviceConnection(fmbtgti.GUITestConnection):
     TizenDeviceConnection copies _tizenAgent to Tizen device,
     and runs & communicates with it via sdb shell.
     """
-    def __init__(self, serialNumber=None, debugAgentFile=None):
-        if serialNumber == None: self._serialNumber = self.recvSerialNumber()
-        else: self._serialNumber = serialNumber
-
+    def __init__(self, serialNumber=None, loginCommand=None, debugAgentFile=None):
+        if loginCommand == None:
+            if serialNumber == None:
+                self._serialNumber = self.recvSerialNumber()
+            else:
+                self._serialNumber = serialNumber
+            self._loginCommand = None
+        else:
+            if serialNumber == None:
+                self._serialNumber = "unknown"
+            else:
+                self._serialNumber = serialNumber
+            if isinstance(loginCommand, str):
+                self._loginCommand = shlex.split(loginCommand)
+            else:
+                self._loginCommand = loginCommand
+        self._useSdb = (self._loginCommand == None)
+        self._useSsh = not self._useSdb
         self._sdbShell = None
         self._debugAgentFile = debugAgentFile
         self.open()
@@ -490,44 +517,72 @@ class TizenDeviceConnection(fmbtgti.GUITestConnection):
         self.close()
 
     def open(self):
-        if self._serialNumber == "unknown":
+        if self._serialNumber == "unknown" and self._useSdb:
             raise TizenDeviceNotFoundError("Tizen device not found.")
 
         self.close()
 
-        agentFilename = os.path.join(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))),
-                                     "fmbttizen-agent.py")
-        agentRemoteFilename = "/tmp/fmbttizen-agent.py"
+        agentFilename = os.path.join(
+            os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))),
+            "fmbttizen-agent.py")
+        remoteUploadPath = "/tmp"
+        agentRemoteFilename = remoteUploadPath + "/fmbttizen-agent.py"
+        uploadFiles = [(agentFilename,
+                        agentRemoteFilename),
+                       (os.path.join(os.path.dirname(agentFilename),
+                                     "fmbtuinput.py"),
+                        remoteUploadPath + "/fmbtuinput.py")]
 
-        uploadCmd = ["sdb", "-s", self._serialNumber, "push", agentFilename, agentRemoteFilename]
+        # Upload fmbttizen-agent to the device
+        if self._useSdb:
+            for src, dst in uploadFiles:
+                uploadCmd = ["sdb", "-s", self._serialNumber, "push", src, dst]
+                status, out, err = _run(uploadCmd, range(256))
+                if status == 127:
+                    raise TizenConnectionError('Executing "sdb -s %s push" failed. Check your Tizen SDK installation.' % (self._serialNumber,))
+                elif status != 0:
+                    if "device not found" in err:
+                        raise TizenDeviceNotFoundError('Tizen device "%s" not found.' % (self._serialNumber,))
+                    else:
+                        raise TizenConnectionError('Executing "%s" failed: %s' % (' '.join(uploadCmd), err + " " + out))
+        else: # using SSH
+            for src, dst in uploadFiles:
+                uploadCmd = self._loginCommand + ["cat", ">", dst]
+                p = subprocess.Popen(uploadCmd, shell=False, stdin=subprocess.PIPE)
+                p.stdin.write(file(src).read())
+                p.stdin.close()
+                p.wait()
 
-        status, out, err = _run(uploadCmd, range(256))
-        if status == 127:
-            raise TizenConnectionError('Executing "sdb -s %s push" failed. Check your Tizen SDK installation.' % (self._serialNumber,))
-        elif status != 0:
-            if "device not found" in err:
-                raise TizenDeviceNotFoundError('Tizen device "%s" not found.' % (self._serialNumber,))
-            else:
-                raise TizenConnectionError('Executing "%s" failed: %s' % (' '.join(uploadCmd), err + " " + out))
+        # Launch agent, create persistent connection to it
+        if self._useSdb:
+            remoteShellCmd = ["sdb", "-s", self._serialNumber, "shell"]
+        else: # using SSH
+            remoteShellCmd = self._loginCommand + ["python", agentRemoteFilename]
 
         try:
-            self._sdbShell = subprocess.Popen(["sdb", "-s", self._serialNumber, "shell"],
+            self._sdbShell = subprocess.Popen(remoteShellCmd,
                                               shell=False,
                                               stdin=subprocess.PIPE,
                                               stdout=subprocess.PIPE,
                                               stderr=subprocess.PIPE,
                                               close_fds=True)
         except OSError, msg:
-            raise TizenConnectionError('Executing "sdb -s %s shell" failed. Check your Tizen SDK installation.' % (self._serialNumber,))
+            raise TizenConnectionError('Executing "%s" failed.' % (
+                " ".join(remoteShellCmd),))
+
         _g_sdbProcesses.add(self._sdbShell)
         self._sdbShellErrQueue = Queue.Queue()
         thread.start_new_thread(_fileToQueue, (self._sdbShell.stderr, self._sdbShellErrQueue))
 
-        self._sdbShell.stdin.write("\r")
-        try:
-            ok, version = self._agentCmd("python %s; exit" % (agentRemoteFilename,))
-        except IOError:
-            raise TizenConnectionError('Connecting to a Tizen device/emulator with "sdb -s %s shell" failed.' % (self._serialNumber,))
+        if self._useSdb:
+            self._sdbShell.stdin.write("\r")
+            try:
+                ok, version = self._agentCmd("python %s; exit" % (agentRemoteFilename,))
+            except IOError:
+                raise TizenConnectionError('Connecting to a Tizen device/emulator with "sdb -s %s shell" failed.' % (self._serialNumber,))
+        else: # using SSH
+            ok, version = self._agentCmd("")
+            pass # agent already started by remoteShellCmd
         return ok
 
     def reportErrorsInQueue(self):
@@ -577,9 +632,14 @@ class TizenDeviceConnection(fmbtgti.GUITestConnection):
     def _agentCmd(self, command, retry=3):
         if self._sdbShell == None: return False, "disconnected"
         if self._debugAgentFile: self._debugAgentFile.write(">0 %s\n" % (command,))
+        if self._useSdb:
+            eol = "\r"
+        else:
+            eol = "\n"
         try:
-            self._sdbShell.stdin.write("%s\r" % (command,))
-            self._sdbShell.stdin.flush()
+            if len(command) > 0:
+                self._sdbShell.stdin.write("%s%s" % (command, eol))
+                self._sdbShell.stdin.flush()
         except IOError, msg:
             if retry > 0:
                 time.sleep(.2)
@@ -628,29 +688,34 @@ class TizenDeviceConnection(fmbtgti.GUITestConnection):
         return status[1]
 
     def recvScreenshot(self, filename, blankFrameRetry=3):
-        if blankFrameRetry > 5:
+        if blankFrameRetry > 2:
             rv, img = self._agentCmd("ss")
         else:
-            rv, img = self._agentCmd("ss R") # resetXConnection
+            rv, img = self._agentCmd("ss R") # retry
         if rv == False:
             return False
-        try:
-            header, data = zlib.decompress(img).split('\n',1)
-            width, height, depth, bpp = [int(n) for n in header.split()[1:]]
-        except Exception, e:
-            raise TizenConnectionError("Corrupted screenshot data: %s" % (e,))
-        if len(data) != width * height * 4:
-            raise FMBTTizenError("Image data size mismatch.")
+        if img.startswith("FMBTRAWX11"):
+            try:
+                header, zdata = img.split('\n', 1)
+                width, height, depth, bpp = [int(n) for n in header.split()[1:]]
+                data = zlib.decompress(zdata)
+            except Exception, e:
+                raise TizenConnectionError("Corrupted screenshot data: %s" % (e,))
 
-        if fmbtgti.eye4graphics.bgrx2rgb(data, width, height) == 0 and blankFrameRetry > 0:
-            time.sleep(0.5)
-            return self.recvScreenshot(filename, blankFrameRetry - 1)
+            if len(data) != width * height * 4:
+                raise FMBTTizenError("Image data size mismatch.")
 
-        # TODO: use libimagemagick directly to save data to png?
-        ppm_header = "P6\n%d %d\n%d\n" % (width, height, 255)
-        f = file(filename + ".ppm", "w").write(ppm_header + data[:width*height*3])
-        _run(["convert", filename + ".ppm", filename], expectedExitStatus=0)
-        os.remove("%s.ppm" % (filename,))
+            if fmbtgti.eye4graphics.bgrx2rgb(data, width, height) == 0 and blankFrameRetry > 0:
+                time.sleep(0.5)
+                return self.recvScreenshot(filename, blankFrameRetry - 1)
+
+            # TODO: use libimagemagick directly to save data to png?
+            ppm_header = "P6\n%d %d\n%d\n" % (width, height, 255)
+            f = file(filename + ".ppm", "w").write(ppm_header + data[:width*height*3])
+            _run(["convert", filename + ".ppm", filename], expectedExitStatus=0)
+            os.remove("%s.ppm" % (filename,))
+        else:
+            file(filename, "w").write(img)
         return True
 
     def recvSerialNumber(self):
